requirements.txt
Django == 1.8.6
django_filter == 0.9.2
requests == 2.7.0
djangorestframework == 3.4.4
setuptools == 1.1.6
Pillow == 3.3.1
django_filters == 0.2.1
django-storages==1.5.0
boto==2.36.0
